added @swagger docs for routes
